requirements.txt
fastapi
uvicorn
tensorflow-cpu==2.20.0
numpy
underthesea
python-multipart
